Okay, back to normal; use conda instead of pip
It looks like using pip is not a good idea for dependencies.
I now think downloading from github and using an identical conda
environment is a much better solution. Each new distribution just needs
to include the state of the environment it was tested in and it's good
to go.