Discovered iraf input limit
In nifsBaselineCalibration we discovered that iraf tasks
can only accept input strings of ~99 characters. The
runtimeData directory name is inside a python package and
has a fairly long name; this can cause a crash.

This commit shortens data file path names as a TEMPORARY fix. If
this comes up as an issue later more effort will need to be invested
to fix it properly.