Fixed minor sort issue
Absolute P and Q offsets are used to identify sky frames; this causes
issues when a target is not located at a P and Q of zero. I implemented
a check that pauses Nifty if a skyFrameList and NO tellist (telluric
frame list) exists and gets them to fix things in the appropriate
directory.